Introduce option 'usebitsets'
When creating a TAP::DOM with new(usebitsets => 1)
then all the 'has_*' and 'is_*' attributes are stored
in a common bitset entry 'is_has' with their
respective bits set to 1.

This reduces the size of a TAP::DOM remarkably
(by ~40% for a large TAP-DOM) and is meant as
an optimization option for memory constrained
problems.