to_tap() works with unblessed data
A TAP::DOM data structure might not contain
the bless information, for example after being
stored in a language-agnostic database.

To call the ->to_tap() method you can bless
such a data structure to TAP::DOM but it should
not be necessary to also bless the elements
inside. Therefore we don't use accessors but
directly the hash elements when traversing the
data structure in to_tap().