fix kv_data vs. disable_global_kv_data
When the option 'disable_global_kv_data' was set
then some key:value data ended up nowhere:

 1. Due to that option it was not put into the
    global document_data.
 2. There was a condition to put it into an item's
    kv_data only when it is a test line, i.e.,
    an 'ok' or 'not ok' line.

However, the traditional way (in some systems, like
Tapper) was to embed such global data right after
the plan. So it wasn't a test line and therefore
not added.

Now we allow the parent to be a plan, too.

We still restrict it with these two parents
'is_test' and 'is_plan' to avoid the kv_data pop up
in unexpected places. Now they appear in structurally
the same place (or "data path"). A counter example
would be a TAP document that immediately starts with

  # Test-key: value

lines right at the beginning without a parent ok line
and the lazy plan coming at the end of the document.
This key:value data is basically still lost when
option 'disable_global_kv_data' is set.