new option 'put_dangling_kv_data_under_lazy_plan'
If the option 'disable_global_kv_data' is used
and key:value lines appear before a plan, those
otherwise dangling/lost lines can be put to a
later appearing plan by setting the new option

  put_dangling_kv_data_under_lazy_plan => 1

This must be requested explcititely because the
data are reordered to appear in another place
(near the last plan line) than they were in the
document (at the start) and this might be
surprising.

Also it's only needed with 'disable_global_kv_data'
because the traditional backwards-compatible place
puts them under a global hash 'document_data'.