new option 'dontignorelines'
HIGHLY EXPERIMENTAL FEATURE!

The option 'dontignorelines' is basically a whitelist
to the blacklist defined with 'ignorelines' so that
you can define exceptions from a broader ignore regex.

In combination with the 'preprocess_ignorelines' option
it is internally converted into a zero-width negative-lookahead
condition and prepended before the 'ignorelines' regex. That
is what makes it so risky/experimental.