introduce TAP::DOM-specific pragma +tapdom_error
pragma +tapdom_error - assign this line a severity of 5

You can for instance append this pragma to the TAP
document during post-processing to express an out-of-band
error situation without interfering with the existing
test lines and plan.

Typical situations can be a an error from "prove" or
other TAP processor, and you want to ensure this problem
does not get lost when storing the document in a database.