noempty_tap - option also replaces empty files
So far only the 'tap' option (i.e. a TAP string) was
taken into consideration to be replaced with an error
document when empty.

Now it also does that when the 'source' points to an
empty file. It set the option 'tap' to the replacement
document and undefines the 'source'.