noempty_tap: also catch no-tap situations
So far the noempty_tap logic only catched real empty
files, either plain TAP or as part of TAP::Archives.

Now it also catches the situation that it's not empty but
does not contain actual TAP lines, which can lead to similar
issues.

A not yet understood fact is that pragma lines sometimes seem
to be of type 'unknown' which is relevant to realize the new
situation against the already handled empty situation.

Also note that when you set 'preprocess_tap' this issue could
be hidden when all those non-TAP lines are already filtered out
so that it actually generates the 'pragma +tapdom_error' with
the alread supported situation comment '# document was empty'.