new option 'ignoreunknown'
By default non-TAP lines are still part of the TAP::DOM (with
is_unknown=1 and most other entry fields set to undef.

If you mix a lot of non-TAP lines with actual TAP lines then this can
lead to a huge TAP::DOM data structure.

With this option set to 1 the unknown lines are skipped.